requirements.txt
autopep8==1.5.4
Django==3.1
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
flake8==3.8.3
Pillow==7.2.0
stripe==2.50.0
psycopg2==2.8.5
django-allauth==0.42.0
